Moving More Complicated Routes into Handlers
